Use real components as a demo data
