+++ browser/ui/webui/brave_webui_source.cc
@@ -64,18 +64,6 @@ void CustomizeWebUIHTMLSource(const std::string &name,
         { "6c337c63662ee0ba4e57f6f8156d69ce.svg", IDR_BRAVE_NEW_TAB_IMG2 },
         // New tab Backgrounds
 #if !defined(OS_ANDROID)
-        { "ntp-dummy-brandedwallpaper/wallpaper-0.jpg",
-            IDR_BRAVE_NEW_TAB_DUMMY_BRANDED_WALLPAPER_BG_1 },
-        { "ntp-dummy-brandedwallpaper/wallpaper-1.jpg",
-            IDR_BRAVE_NEW_TAB_DUMMY_BRANDED_WALLPAPER_BG_2 },
-        { "ntp-dummy-brandedwallpaper/wallpaper-2.jpg",
-            IDR_BRAVE_NEW_TAB_DUMMY_BRANDED_WALLPAPER_BG_3 },
-        { "ntp-dummy-brandedwallpaper/logo.png",
-            IDR_BRAVE_NEW_TAB_DUMMY_BRANDED_WALLPAPER_LOGO },
-        { "ntp-dummy-brandedwallpaper/brave.png",
-            IDR_BRAVE_NEW_TAB_DUMMY_BRANDED_WALLPAPER_BRAVE_ICON },
-        { "ntp-dummy-brandedwallpaper/bat.png",
-            IDR_BRAVE_NEW_TAB_DUMMY_BRANDED_WALLPAPER_BAT_ICON },
         { "anders-jilden.webp", IDR_BRAVE_NEW_TAB_BACKGROUND1 },
         { "andreas-gucklhorn.webp", IDR_BRAVE_NEW_TAB_BACKGROUND2 },
         { "andy-mai.webp", IDR_BRAVE_NEW_TAB_BACKGROUND3 },
