+++ components/ntp_background_images/browser/view_counter_service_unittest.cc
@@ -3,6 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+#include <memory>
 #include <string>
 
 #include "base/files/file_path.h"
@@ -21,6 +22,30 @@
 #include "testing/gtest/include/gtest/gtest.h"
 
 namespace ntp_background_images {
+
+std::unique_ptr<NTPBackgroundImagesData> GetDemoWallpaper(bool super_referrer) {
+  auto demo = std::make_unique<NTPBackgroundImagesData>();
+  demo->url_prefix = "chrome://newtab/ntp-dummy-brandedwallpaper/";
+  demo->backgrounds = {
+      { base::FilePath(FILE_PATH_LITERAL("wallpaper1.jpg")), { 3988, 2049 } },
+      { base::FilePath(FILE_PATH_LITERAL("wallpaper2.jpg")), { 5233, 3464 } },
+      { base::FilePath(FILE_PATH_LITERAL("wallpaper3.jpg")), {  0, 0 } },
+  };
+  demo->logo_alt_text = "Technikke: For music lovers.";
+  demo->logo_company_name = "Technikke";
+  demo->logo_destination_url = "https://brave.com";
+
+  if (super_referrer) {
+    demo->top_sites = {
+      { "Brave", "https://brave.com", "brave.png",
+        base::FilePath(FILE_PATH_LITERAL("brave.png")) },
+     { "BAT", "https://basicattentiontoken.org/", "bat.png",
+        base::FilePath(FILE_PATH_LITERAL("bat.png")) },
+    };
+  }
+
+  return demo;
+}
 
 class NTPBackgroundImagesViewCounterTest : public testing::Test {
  public:
